pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "order-sync"
version = "0.1.0"
description = "Sync orders per user into Excel workbooks"
requires-python = ">=3.9"
dependencies = [
    "openpyxl>=3.1.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.116.0",
    "google-auth>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "pymongo[srv]>=4.6.0",
]

[project.scripts]
order-sync = "order_sync.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
